refactor: move multipart boundary generation to BrowserFingerprint
